test: move to unit dir
